requirements.txt
imgkit==1.2.3
Pillow==9.4.0
six==1.16.0
zpl==0.1.10
